Only build the aim and text debuggers if this build is not a release
These have been moved to their own files with go:build constraint
comments in them so that they are only built when the including the
release tag, like this:

    go build -tags release .

By default (e.g. in development) this is not the case, so the debugging
will be present here, but there is now a convenient way to consistently
turn it off for the production release that other people play.
